requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
alembic
langchain
langchain-google-genai
google-generativeai
google-generativeai==0.8.2
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow>=10.1.0
whisper==1.1.10
youtube-dl==2021.12.17
beautifulsoup4==4.12.2
requests==2.31.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
langchain>=0.1.0
langchain-openai>=0.1.0
chromadb>=0.5.0
numpy>=1.26.4
PyMuPDF>=1.22.0
youtube-transcript-api>=0.6.1
reportlab>=4.0.0
qrcode>=7.4.2
passlib[bcrypt]
bcrypt==4.0.1
huggingface_hub>=0.23.0
slowapi>=0.1.9
python-magic>=0.4.27
